Split mission tests
Also, refactor the tests so they don't directly use the json library
and to use common function to get the mission list
